requirements.txt
# Core dependencies
# Note: Using typer without [all] to avoid rich help formatting issues
typer==0.7.0
click==8.1.3
rich==12.6.0
commonmark==0.9.1
pandas==2.2.0
numpy==1.26.3
pyyaml==6.0.1
requests==2.31.0
ccxt==4.2.25
python-dateutil==2.8.2
pyarrow==15.0.0

# Optional for future ML
scikit-learn==1.4.0
